Move `improv.cpp` and `improv.h` to own integration folder
